refactor: Simplify ExpandedConfig class
Removed the enum `ExpandedScheme` and changed the type of the `$scheme` property in the `ExpandedConfig` class from `ExpandedScheme` to `string`. Also removed the `encryptionKey` property and its references from the class. Simplified the `create` method accordingly. Added a `toObject` method to convert an instance of `ExpandedConfig` to a JSON string.