feat: Add Logging trait for error logging
Added a new Logging trait in the src/Traits directory to facilitate error logging. This trait provides a log method to log messages to a file named errors.log in the user's home directory under the .libsql-php directory. The log method appends the current date and time along with the message to the log file.