feat: Implement logging in LibsqlError class
Added logging functionality to the LibsqlError class located at src/Utils/Exceptions/LibsqlError.php. This allows any error messages thrown by the LibsqlError class to be logged to a file for debugging and error tracking purposes. The log method is called with the error message upon instantiation of the LibsqlError class, ensuring that all errors are logged.